% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drvid-package.R
\docType{package}
\name{drvid-package}
\alias{drvid}
\alias{drvid-package}
\title{drvid: Access to 'DVID' API}
\description{
API access to the 'DVID' (Distributed, Versioned, Image-oriented 
    Dataservice).
}
\section{Package environment variables}{


  The following environment variables may be set e.g. in your
  \code{\link{Renviron}} file in order to specify a default DVID server. See
  \code{\link{dv_conn}} for more information about specifying DVID login
  details.

  \itemize{

  \item drvid.server

  \item drvid.node

  }
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferis/drvid}
  \item Report bugs at \url{https://github.com/jefferis/drvid/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (0000-0002-0587-9355)

}
\keyword{internal}
