% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{dv_get_annotations}
\alias{dv_get_annotations}
\title{Retrieve annotations for bodyids}
\usage{
dv_get_annotations(bodyids, conn = NULL, ...)
}
\arguments{
\item{conn}{optional DVID connection object (see \code{\link{dv_conn}})}

\item{...}{Additional arguments passed to dv_fetch}
}
\value{
a dataframe with information on names, annotations and the users that  gave them for the given bodyids
}
\description{
Retrieve annotations for a bodyid or vector of bodyids
}
