% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{dv_get_segmentation_info}
\alias{dv_get_segmentation_info}
\title{Get information about a segmentation}
\usage{
dv_get_segmentation_info(conn = NULL, ...)
}
\arguments{
\item{conn}{optional DVID connection object (see \code{\link{dv_conn}})}

\item{...}{Additional arguments passed to dv_fetch}
}
\value{
a list of information, possibly including: BlockSize, VoxelSize, VoxelUnits, MinPoint, MaxPoint, MinIndex, MaxIndex,
Background, MaxLabel, MaxRepoLabel, IndexedLabels, MaxDownresLevel, DataUUID, RepoUUID, Compression, etc.
}
\description{
Get information about the segmentation hosted on a DVID server
}
