% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{dv_get_voxels}
\alias{dv_get_voxels}
\alias{dv_get_mesh}
\title{Get a mesh or voxel data for a given bodyid}
\usage{
dv_get_voxels(bodyid, scale = 4, conn = NULL, ...)

dv_get_mesh(bodyid, scale = 4, conn = NULL,
  use.surface.voxels = TRUE, k = 10, ...)
}
\arguments{
\item{bodyid}{a body ID for a neuron or segmentation hosted on a DVID server}

\item{scale}{Resolution of sparse volume starting with 0 where each level
beyond 0 has 1/2 resolution of previous level. "coarse" will
return the volume in block coordinates.}

\item{conn}{optional DVID connection object (see \code{\link{dv_conn}})}

\item{...}{Additional arguments passed to dv_fetch}

\item{use.surface.voxels}{if TRUE, surface voxels as estimated as points that do not have at least k other points within 2*voxelsize of them}

\item{k}{see use.surface.voxels}
}
\value{
a mesh3
}
\description{
Get a mesh or voxel data for a given bodyid, from a DVID server.
Meshes are made via alphashapes, using the function \code{alphashape3d::ashape3d}, which will work poorly is scale is too low, and slowly is scale is too high.
}
