% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{read.neuron.dvid}
\alias{read.neuron.dvid}
\alias{read.neurons.dvid}
\title{Read a single neuron in skeleton format from DVID server}
\usage{
read.neuron.dvid(x, conn = NULL, ...)

read.neurons.dvid(x, conn = NULL, OmitFailures = NA, df = NULL, ...)
}
\arguments{
\item{x}{body id}

\item{conn}{optional DVID connection object (see \code{\link{dv_conn}})}

\item{...}{Additional arguments passed to nat::read.neuron}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in nlapply stopping with
an error message the moment there is an error. For other values, see
details.}

\item{df}{Optional data frame of metadata with one row per neuron}
}
\value{
a \code{\link{neuronlist}} object or a \code{\link{neuron}} for
  \code{read.neuron.dvid}
}
\description{
Read a single neuron in skeleton format from DVID server
}
\examples{
\dontrun{
nl=read.neuron.dvid(c(12345, 45678))
}
}
