% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{dv_conn}
\alias{dv_conn}
\title{Make a connection to DVID server}
\usage{
dv_conn(server = Sys.getenv("drvid.server"),
  node = Sys.getenv("drvid.node"), conn = NULL, ...)
}
\arguments{
\item{server}{DVID server (with optional port specification)}

\item{node}{DVID node}

\item{conn}{A pre-existing connection object}

\item{...}{Named options / headers to be set in the curl handle (see
\code{\link[curl]{handle}} and \code{\link[curl]{curl_options}} for
details)}
}
\value{
An list of class \code{dv_conn}
}
\description{
Make a connection to DVID server
}
\examples{
\dontrun{

conn1 <- dv_conn('http://dvid.connectomesrus.com:8900', node='a32b')
conn2 <- dv_conn('http://dvid.connectomesrus.com:8900', node='a42c')

}
}
\seealso{
\code{\link[curl]{handle}}, \code{\link[curl]{curl_options}}
}
